requirements.txt
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.2
fonttools==4.38.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
